third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-column-rule-expected.txt

Tests that the -webkit-column-rule shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('-webkit-column-rule') is "1px dotted rgb(0, 0, 0)"
PASS computedStyle.getPropertyValue('-webkit-column-rule') is "0px none rgb(255, 0, 0)"
PASS computedStyle.getPropertyValue('-webkit-column-rule') is "3px dashed rgb(255, 0, 0)"
PASS computedStyle.getPropertyValue('-webkit-column-rule') is "0px none rgb(0, 0, 255)"
PASS computedStyle.getPropertyValue('-webkit-column-rule') is "0px hidden rgb(0, 0, 255)"
PASS successfullyParsed is true

TEST COMPLETE
